qcacld-3.0: Fix improper type casting and return status
Propagate from qcacld-2.0 to qcacld-3.0

In rrmProcessBeaconReportReq(), the return value is not the same
enum type with the routine definition. In dfs_process_radarevent(),
2 variables type are not same in a comparison. Add fix to correct
it.

CRs-Fixed: 979671

+++ core/sap/dfs/src/dfs_process_radarevent.c
@@ -721,7 +721,8 @@ int dfs_process_radarevent(struct ath_dfs *dfs,
 						 */
 
 						if ((deltaT >
-						     (dfs->dfs_pri_multiplier *
+						     ((u_int64_t)dfs->
+						       dfs_pri_multiplier *
 						      rf->rf_maxpri))
 						    || (deltaT <
 							rf->rf_minpri)) {
